Improvements to testability of SyncEngine.
